Add files via upload
This project involves real-time pose estimation using TensorFlow.js and the Teachable Machine Pose model. Using the webcam, this application allows the user's body movements to be recorded daily, reflecting the display of predictions in the order of poses, as well as controlling a moving box with the arrow keys.